% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{autoBandPassFilter}
\alias{autoBandPassFilter}
\title{Applies automatic bandpass filter to AM sound}
\usage{
autoBandPassFilter(wave, bps = 2)
}
\arguments{
\item{wave}{an R object}

\item{bps}{the bandpass size relative to \deqn{\sigma}{sigma}}
}
\value{
a bandpass filter wave of the same type as \code{wave}
}
\description{
This function assumes that the wave is a recording featuring mainly 
a single source, amplitude modulated song and tries to find and apply automatic bandpass filter
in order to reduce noise.
}
\details{
The \code{bps} arguments refers to the stringency of the filter.
It represents the width of the bandpass filter relative to
the standard deviation of the PDF, in frequency range.
In the original publication \code{bps} = 2 (i.e. \deqn{2\sigma}{2 sigmas}, see eq. 3).
The present implementation allow to alter this parameter. 
Higher values will result in less broad/specific filter.
}
\note{
todo ref (Dietrich et al., 2004)
}

