% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{standardiseWave}
\alias{standardiseWave}
\title{Standardise a Wave object for package consistency}
\usage{
standardiseWave(wave, f = 44100, stereo = FALSE, bit = 1)
}
\arguments{
\item{wave}{an R object or the path of a wave file}

\item{f}{the expected target frequency}

\item{stereo}{wether the expected object is stereo (else it is mono)}

\item{bit}{the bit depth}
}
\value{
a \code{Wave}  object that may have been altered to match package standard representation
}
\description{
This function defines the internal data representation used in this package.
It is interanlly called in most other functions.
}

